requirements.txt
docutils
vobject
whoosh
Markdown

http://www.djangoproject.com/download/1.1.1/tarball/

# Django third party apps
django-shorturls
django-extensions
-e git+git://github.com/toastdriven/django-haystack.git#egg=django-haystack
-e git+git://github.com/svetlyak40wt/django-perfect404.git#egg=Django-Perfect404
-e git+git://github.com/leah/django-elsewhere.git#egg=Django-Elsewhere
South
http://bitbucket.org/jespern/django-piston/downloads/django-piston-0.2.2.tar.gz

